0316 오전 커밋

+++ WinformApp/WinFormAdvancedBank/AddressInfoApp/FrmMain.cs
@@ -119,26 +119,32 @@ namespace AddressInfoApp
                 MessageBox.Show("삭제하려면 데이터를 선택하세요");
                 return;
             }
-            using (SqlConnection conn = new SqlConnection(connString))
+
+            if (MessageBox.Show("삭제하시겠습니까?", "삭제", MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+                == DialogResult.Yes)
             {
-                if (conn.State == ConnectionState.Closed)
-                    conn.Open();
-
-                string query = $"DELETE FROM Address WHERE idx = {result}";
-
-                SqlCommand cmd = new SqlCommand(query, conn);
-                if (cmd.ExecuteNonQuery() == 1)
+                using (SqlConnection conn = new SqlConnection(connString))
                 {
-                    MessageBox.Show("삭제 성공!");
+                    if (conn.State == ConnectionState.Closed)
+                        conn.Open();
+
+                    string query = $"DELETE FROM Address WHERE idx = {result}";
+
+                    SqlCommand cmd = new SqlCommand(query, conn);
+                    if (cmd.ExecuteNonQuery() == 1)
+                    {
+                        MessageBox.Show("삭제 성공!");
+                    }
+                    else
+                    {
+                        MessageBox.Show("삭제 실패!");
+                    }
                 }
-                else
-                {
-                    MessageBox.Show("삭제 실패!");
-                }
+
+                RefreshData();
+                ClearInput();
             }
-
-            RefreshData();
-            ClearInput();
+            
         }
 
         private void TxtFullName_KeyPress(object sender, KeyPressEventArgs e)
